Add chart and rollout fasit feature on commit to main
In order to implement proper CD for elector we need a way to propagate the newly built image to naiserator.
This feature will install a configmap in the cluster which we then can refer to in naiserator.
Combined with config reloading on update, we should get proper CD for elector.